refactor(web): extract ArtifactPlanner sub-components into files
Move the four private sub-components out of the single
artifact-planner.tsx into a components/artifact-planner/ directory:
main-affix-selector, set-configuration, artifact-set-search, and
priority-substats, with artifact-planner.tsx reduced to a thin
composition wrapper and an index barrel preserving the
@/components/artifact-planner import path.

The searchable-select (ArtifactSetSearch) and reorderable-list
(PrioritySubstats) patterns each have exactly one consumer, so they
stay concrete rather than becoming generic SearchableSelect /
ReorderableList utilities; extract when a second use appears.

Closes #610
