+++ apps/web/src/components/artifact-planner/index.ts
@@ -1,0 +1,4 @@
+// SPDX-FileCopyrightText: 2026 Alex Brandt <alunduil@gmail.com>
+// SPDX-License-Identifier: MIT
+
+export { ArtifactPlanner } from './artifact-planner';
